# docs/OpenAPI.yaml
openapi: 3.0.3
info:
  title: PawPals API
  version: 1.0.0
  description: |
    REST API for the PawPals mobile app. The frontend must not access the database
    or Supabase directly; all data access goes through this API layer.
    Provides CRUD operations for profiles, dogs, matches, playdates, ratings, and messages.
servers:
  - url: https://api.pawpals.example.com/v1
    description: Production server (example)
  - url: http://localhost:4000/v1
    description: Local dev server
...
paths:
...
  /dogs:
    get:
      tags: [Dogs]
      summary: List dogs
      operationId: listDogs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: ownerId
          schema:
            type: string
            format: uuid
          required: false
          description: Filter by owner (auth.users.id)
      responses:
        '200':
          description: List of dogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dog'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Dogs]
      summary: Create a dog
      operationId: createDog
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DogCreate'
      responses:
        '201':
          description: Created dog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /dogs/{dogId}:
    get:
      tags: [Dogs]
      summary: Get a dog by ID
      operationId: getDogById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DogIdParam'
      responses:
        '200':
          description: Dog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Dog not found
    patch:
      tags: [Dogs]
      summary: Update a dog
      operationId: updateDog
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DogIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DogUpdate'
      responses:
        '200':
          description: Updated dog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Dog not found
    delete:
      tags: [Dogs]
      summary: Delete a dog
      operationId: deleteDog
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DogIdParam'
      responses:
        '204':
          description: Dog deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /matches:
    get:
      tags: [Matches]
      summary: List matches for current user
      description: Returns matches where the user owns either dog in the pair.
      operationId: listMatches
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Matches]
      summary: Create a match between two dogs
      operationId: createMatch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchCreate'
      responses:
        '201':
          description: Created match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /matches/{matchId}:
    get:
      tags: [Matches]
      summary: Get a match by ID
      operationId: getMatchById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MatchIdParam'
      responses:
        '200':
          description: Match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Match not found
    patch:
      tags: [Matches]
      summary: Update a match (e.g., status)
      operationId: updateMatch
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MatchIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchUpdate'
      responses:
        '200':
          description: Updated match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Match not found
    delete:
      tags: [Matches]
      summary: Delete a match
      description: Cascades to playdates, ratings, and messages due to foreign keys.
      operationId: deleteMatch
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MatchIdParam'
      responses:
        '204':
          description: Match deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DogIdParam:
      in: path
      name: dogId
      required: true
      schema:
        type: string
        format: uuid
    MatchIdParam:
      in: path
      name: matchId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized – missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden – user not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Dog:
      type: object
      properties:
        dog_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        breed:
          type: string
        age:
          type: integer
          minimum: 0
        vax_status:
          type: string
          nullable: true
        spay_neuter_status:
          type: string
          nullable: true
        photo_count:
          type: integer
          minimum: 3
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - dog_id
        - user_id
        - name
        - breed
        - age
        - photo_count
        - created_at
        - updated_at
    DogCreate:
      type: object
      properties:
        name:
          type: string
        breed:
          type: string
        age:
          type: integer
          minimum: 0
        vax_status:
          type: string
        spay_neuter_status:
          type: string
        photo_count:
          type: integer
          minimum: 3
      required:
        - name
        - breed
        - age
        - photo_count
    DogUpdate:
      type: object
      properties:
        name:
          type: string
        breed:
          type: string
        age:
          type: integer
          minimum: 0
        vax_status:
          type: string
        spay_neuter_status:
          type: string
        photo_count:
          type: integer
          minimum: 3
    Match:
      type: object
      properties:
        match_id:
          type: string
          format: uuid
        dog1_id:
          type: string
          format: uuid
        dog2_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        status:
          type: string
          description: Match status (e.g., pending, confirmed, blocked)
      required:
        - match_id
        - dog1_id
        - dog2_id
        - created_at
        - status
    MatchCreate:
      type: object
      properties:
        dog1_id:
          type: string
          format: uuid
        dog2_id:
          type: string
          format: uuid
      required:
        - dog1_id
        - dog2_id
    MatchUpdate:
      type: object
      properties:
        status:
          type: string
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Something went wrong
        details:
          type: object
          additionalProperties: true